+++ com.unity.render-pipelines.core/Runtime/Volume/VolumeParameter.cs
@@ -1447,6 +1447,19 @@ namespace UnityEngine.Rendering
             : base(value, overrideState) { }
 
         // TODO: Texture interpolation
+
+        public override int GetHashCode()
+        {
+            int hash = base.GetHashCode();
+
+            unchecked
+            {
+                if (value != null)
+                    hash = 23 * CoreUtils.GetTextureHash(value);
+            }
+            
+            return hash;
+        }
     }
 
     /// <summary>
@@ -1462,6 +1475,19 @@ namespace UnityEngine.Rendering
         /// <param name="overrideState">The initial override state for the parameter.</param>
         public NoInterpTextureParameter(Texture value, bool overrideState = false)
             : base(value, overrideState) { }
+
+        public override int GetHashCode()
+        {
+            int hash = base.GetHashCode();
+
+            unchecked
+            {
+                if (value != null)
+                    hash = 23 * CoreUtils.GetTextureHash(value);
+            }
+            
+            return hash;
+        }
     }
 
     /// <summary>
@@ -1477,6 +1503,19 @@ namespace UnityEngine.Rendering
         /// <param name="overrideState">The initial override state for the parameter.</param>
         public Texture2DParameter(Texture value, bool overrideState = false)
             : base(value, overrideState) { }
+
+        public override int GetHashCode()
+        {
+            int hash = base.GetHashCode();
+
+            unchecked
+            {
+                if (value != null)
+                    hash = 23 * CoreUtils.GetTextureHash(value);
+            }
+            
+            return hash;
+        }
     }
 
     /// <summary>
@@ -1492,6 +1531,19 @@ namespace UnityEngine.Rendering
         /// <param name="overrideState">The initial override state for the parameter.</param>
         public Texture3DParameter(Texture value, bool overrideState = false)
             : base(value, overrideState) { }
+
+        public override int GetHashCode()
+        {
+            int hash = base.GetHashCode();
+
+            unchecked
+            {
+                if (value != null)
+                    hash = 23 * CoreUtils.GetTextureHash(value);
+            }
+            
+            return hash;
+        }
     }
 
     /// <summary>
@@ -1509,6 +1561,19 @@ namespace UnityEngine.Rendering
             : base(value, overrideState) { }
 
         // TODO: RenderTexture interpolation
+
+        public override int GetHashCode()
+        {
+            int hash = base.GetHashCode();
+
+            unchecked
+            {
+                if (value != null)
+                    hash = 23 * CoreUtils.GetTextureHash(value);
+            }
+            
+            return hash;
+        }
     }
 
     /// <summary>
@@ -1524,6 +1589,19 @@ namespace UnityEngine.Rendering
         /// <param name="overrideState">The initial override state for the parameter.</param>
         public NoInterpRenderTextureParameter(RenderTexture value, bool overrideState = false)
             : base(value, overrideState) { }
+
+        public override int GetHashCode()
+        {
+            int hash = base.GetHashCode();
+
+            unchecked
+            {
+                if (value != null)
+                    hash = 23 * CoreUtils.GetTextureHash(value);
+            }
+            
+            return hash;
+        }
     }
 
     /// <summary>
@@ -1540,6 +1618,19 @@ namespace UnityEngine.Rendering
         public CubemapParameter(Texture value, bool overrideState = false)
             : base(value, overrideState) { }
         // TODO: Cubemap interpolation
+
+        public override int GetHashCode()
+        {
+            int hash = base.GetHashCode();
+
+            unchecked
+            {
+                if (value != null)
+                    hash = 23 * CoreUtils.GetTextureHash(value);
+            }
+            
+            return hash;
+        }
     }
 
     /// <summary>
@@ -1555,6 +1646,19 @@ namespace UnityEngine.Rendering
         /// <param name="overrideState">The initial override state for the parameter.</param>
         public NoInterpCubemapParameter(Cubemap value, bool overrideState = false)
             : base(value, overrideState) { }
+    
+        public override int GetHashCode()
+        {
+            int hash = base.GetHashCode();
+
+            unchecked
+            {
+                if (value != null)
+                    hash = 23 * CoreUtils.GetTextureHash(value);
+            }
+            
+            return hash;
+        }
     }
 
     /// <summary>