+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fixed artifacts on gpu light culling when the frustum near and far are very spread out (case 1386436)
 - Fixed an issue where forced sky update (like PBR sky amortized updated) would not update ambient probe.
 - Fixed a null ref exception in Volume Explorer
+- Fixed static lighting sky update when using an HDRI sky with a render texture in parameter.
 
 ## [12.1.4] - 2021-12-07